+++ src/winpriv.h
@@ -82,8 +82,6 @@ extern void win_init_drop_target(void);
 
 extern wstring wslicon(wchar * params);
 
-extern char * foreground_cwd(void);
-
 extern void win_switch(bool back, bool alternate);
 extern int search_monitors(int * minx, int * miny, HMONITOR lookup_mon, bool get_primary, MONITORINFO *mip);